+++ src/errors/index.ts
@@ -1,1 +1,2 @@
 export { default as ResourceNotFound } from './ResourceNotFound';
+export { default as IncompleteDocument } from './IncompleteDocument';